Separate index.js into server.js and app.js
